+++ pages/cesiumView.js
@@ -12,7 +12,14 @@ window.CesiumView = function() {
     // But the pole doesn't get much sunlight.  STK says 23.444 degrees solar elevation
     // on 20 Jun 2020 12:11:47.754.  But that puts the Sun straight behind the camera, so
     // I moved it a little.  The Sun stays close to that elevation all day.
-    var currentTime = Cesium.JulianDate.fromIso8601('2020-06-20T14:00:00Z');
+    var startTime = Cesium.JulianDate.fromIso8601('2020-06-20T14:00:00Z');
+
+    // Still need to animate time forwards, to enable glTF animations.
+    var clock = new Cesium.Clock({
+        startTime: startTime,
+        currentTime: startTime,
+        shouldAnimate: true
+    });
 
     function resize() {
         var zoomFactor = Cesium.defaultValue(window.devicePixelRatio, 1.0);
@@ -72,6 +79,7 @@ window.CesiumView = function() {
 
         scene.initializeFrame();
         resize();
+        var currentTime = clock.tick();
         scene.render(currentTime);
         Cesium.requestAnimationFrame(startRenderLoop);
     }