+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 2.1.5 - UNRELEASED
 
 * Added animation controls to the ThreeJS preview window. [#86](https://github.com/AnalyticalGraphicsInc/gltf-vscode/pull/86)
+* Fixed an issue where Cesium animations stopped, introduced in the previous version.
 
 ### 2.1.4 - 2018-03-08